+++ src/main/java/savi_ros_java/savi_ros_bdi/ActionTalker.java
@@ -1,4 +1,4 @@
-package com.github.rosjava.savi_ros_java.savi_ros_bdi;
+package rosjava.savi_ros_java.savi_ros_bdi;
 
 import org.ros.concurrent.CancellableLoop;
 import org.ros.namespace.GraphName;
