+++ src/main/java/savi_ros_java/savi_ros_bdi/SAVI_Main.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.github.rosjava.savi_ros_java.savi_ros_bdi;
+package rosjava.savi_ros_java.savi_ros_bdi;
 
 import org.apache.commons.logging.Log;
 import org.ros.message.MessageListener;
@@ -26,19 +26,35 @@ import org.ros.node.topic.Subscriber;
 
 import jason.asSyntax.*;
 
-public class PerceptionListener implements Runnable {
+/**
+ * A simple {@link Subscriber} {@link NodeMain}.
+ */
+public class SAVI_Main extends AbstractNodeMain {
 
-    // The node for passing messages to
-    private ConnectedNode connectedNode;
+    private SaviAgentArch agent;
 
-    private PerceptionListener() { }
-
-    public PerceptionListener(final ConnectedNode connectedNode) {
-        // Set the connectedNode parameter
-        this.connectedNode = connectedNode;
+    /**
+     * Provide name of this node when requested.
+     * @return
+     */
+    @Override
+    public GraphName getDefaultNodeName() {
+        return GraphName.of("rosjava/listener");
     }
 
-    public void run() {
+
+    /**
+     * Start method for the node (can think of this as the 'main()' method.
+     * @param connectedNode
+     */
+    @Override
+    public void onStart(ConnectedNode connectedNode) {
+
+        // Initialize the agent
+        this.agent = new SaviAgentArch(connectedNode);
+        this.agent.startAgents();
+
+        /*
         final Log log = connectedNode.getLog();
         Subscriber<std_msgs.String> subscriber = connectedNode.newSubscriber("perceptions", std_msgs.String._TYPE);
         subscriber.addMessageListener(new MessageListener<std_msgs.String>() {
@@ -52,5 +68,6 @@ public class PerceptionListener implements Runnable {
                 agentState.setPerceptions(rxLiteral.toString());
             }
         });
+        */
     }
 }
