+++ src/main/java/savi_ros_java/savi_ros_bdi/SaviAgent.java
@@ -1,4 +1,4 @@
-package com.github.rosjava.savi_ros_java.savi_ros_bdi;
+package rosjava.savi_ros_java.savi_ros_bdi;
 
 import jason.architecture.AgArch;
 import jason.asSemantics.*;
